gendoor.c FUN_80156710, correct_link

+++ include/gendoor.h
@@ -11,9 +11,9 @@
 extern u8 *link_init; /* not sure if this should be here? since non-gendoor things use links? */
 
 void deactivate_ship_links(void);
-s32 linkListHoldsAGendoor(Obj *obj);
-s32 FUN_8015666c(Obj *obj);
-s32 FUN_80156710(Obj *obj);
+s16 linkListHoldsAGendoor(Obj *obj);
+s16 FUN_8015666c(Obj *obj);
+s16 FUN_80156710(Obj *obj);
 void correct_gendoor_link(void);
 void suppressFromLinkList(Obj *obj);
 void correct_link(void);
